movies and actors page and add minor changes
