+++ frontend/src/pages/index.ts
@@ -1,3 +1,5 @@
 export { default as Home } from "./Home";
 export { default as Auth } from "./Auth";
 export { default as Page404 } from "./Page404";
+export { default as Movies } from "./Movies";
+export { default as Actors } from "./Actors";
