+++ frontend/src/styles/App.scss
@@ -108,3 +108,37 @@ a {
         text-decoration: underline;
     }
 }
+
+.sectionContainer {
+    padding: var(--spacing-lg) var(--spacing-md);
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    .innerContainer {
+        min-width: var(--site-max-width);
+        max-width: var(--site-max-width);
+
+        .sectionTop {
+            display: flex;
+            justify-content: space-between;
+            align-items: flex-start;
+            .seeAll {
+                text-transform: uppercase;
+                font-size: var(--fz-md);
+                font-weight: 700;
+                margin-bottom: var(--spacing-xs);
+                color: var(--light-grey);
+            }
+        }
+
+        .container {
+            display: grid;
+            grid-template-columns: repeat(auto-fill, minmax(250px, 0.8fr));
+            grid-gap: var(--spacing-xl);
+            justify-items: center;
+            max-width: var(--site-max-width);
+        }
+    }
+}
